# api/openapi.yaml
openapi: 3.1.0
info:
  title: SuperMonitor API
  version: 0.9.0
  description: Monitoring-only API. It intentionally exposes no model proxy endpoints.
servers:
  - url: /api/v1
security:
  - AdminSession: []
  - BearerAuth: []
paths:
...
  /auth/session:
    post:
      security: []
      operationId: createAuthenticationSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  minLength: 12
                  writeOnly: true
                  description: Administrator password containing at least 12 Unicode characters
                remember:
                  type: boolean
                  default: true
                  description: Keep the browser session for 30 days instead of 12 hours
      responses:
        "200":
          description: HttpOnly administrator session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "401":
          description: Administrator password is invalid
    delete:
      operationId: deleteAuthenticationSession
      responses:
        "200":
          description: Browser administrator session cleared
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
  /auth/connect:
    post:
      security: []
      operationId: connectAuthenticationSession
      description: Browser form login for a user-selected SuperMonitor address. Passwords are exchanged for an HttpOnly session and never placed in a URL.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [address, password]
              properties:
                address: { type: string, format: uri }
                password: { type: string, minLength: 12, writeOnly: true }
                remember: { type: string, enum: ["1"] }
      responses:
        "303":
          description: Redirect to the console or back to the login gate with a non-sensitive error code
...
  /providers/kiro/oauth/callback:
    get:
      security: []
      operationId: completeKiroOAuth
      description: Public one-time OAuth callback protected by the server-held state and PKCE verifier.
      parameters:
        - in: query
          name: state
          required: true
          schema: { type: string }
        - in: query
          name: code
          schema: { type: string }
        - in: query
          name: error
          schema: { type: string }
        - in: query
          name: login_option
          schema: { type: string }
      responses:
        "200":
          description: Kiro authorization completed; the browser may close this page
          content:
            text/html:
              schema: { type: string }
        "400":
          description: Invalid, expired, denied, or failed Kiro authorization callback
          content:
            text/html:
              schema: { type: string }
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: supermonitor_admin
      description: Server-issued HttpOnly administrator session (12 hours or 30 days when remembered)
    BearerAuth:
      type: http
      scheme: bearer
      description: SUPMON_ADMIN_PASSWORD for API automation
  schemas:
    AuthStatus:
      type: object
      required: [required, authenticated]
      properties:
        required: { type: boolean }
        authenticated: { type: boolean }
...